requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
diploma-frontend @ file:///Users/umbarloce/Desktop/Python_django_diploma_dpo/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=37600d2f30a9372e7e0294b9ac7dd8d165d18ddf2139b515f7bdf12bb470db78
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
